Tweaking ID generation code, fixing a couple of bugs
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@6547 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ tools/riddler/riddler-ejb/src/java/edu/cmu/sphinx/tools/riddler/ejb/RiddlerRemote.java
@@ -26,7 +26,7 @@ public interface RiddlerRemote {
      * @param corpusID Corpus from which the models should be trained
      * @return a URI pointing to an RSS feed that will contain the models matching this request
      */
-    public URI trainModelsFromCorpus(int corpusID);
+    public URI trainModelsFromCorpus(long corpusID);
 
     /**
      * Kick off a model training operation that uses multiple Corpora
@@ -41,7 +41,7 @@ public interface RiddlerRemote {
      * @throws RemoteException if a Dictionary with the given metadata already exists
      * @param metadata a map of metadata about the Dictionary
      */
-    public int createDictionary(MetadataWrapper metadata);
+    public long createDictionary(MetadataWrapper metadata);
 
     /**
      * Get a Dictionary matching the provided descriptor
@@ -49,14 +49,14 @@ public interface RiddlerRemote {
      * @throws RemoteException if no Dictionary matching the metadata exists
      * @param metadata metadata for the Dictionary you're trying to retrieve
      */
-    public int getDictionary(MetadataWrapper metadata);
+    public long getDictionary(MetadataWrapper metadata);
 
     /**
      * Fetch the DictionaryDescriptor matching the provided ID
-     * @param dictionaryID a legitimate Dictionary identifier
-     * @return the DictionaryDescriptor provided when the given Dictionary was created
+     * @param dictionaryID an existing Dictionary identifier
+     * @return the metadata provided when the given Dictionary was created
      */
-    public MetadataWrapper getDictionaryMetadata(int dictionaryID);
+    public MetadataWrapper getDictionaryMetadata(long dictionaryID);
 
     /**
      * Create a Pronunciation record (case insensitive).
@@ -67,7 +67,7 @@ public interface RiddlerRemote {
      * @param word the word to be added to the specified dictionary
      * @param pronunciations main and any variant pronunciations of the word
      */
-    public int createPronuncation(int dictionaryID, String word, ArrayList<String> pronunciations);
+    public long createPronuncation(long dictionaryID, String word, ArrayList<String> pronunciations);
 
     /**
      * Check whether a Pronunciation record exists for the given word (case insensitive)
@@ -75,7 +75,7 @@ public interface RiddlerRemote {
      * @param word word that should be queried
      * @return true if the word has at least one pronunciation
      */
-    public boolean hasPronuncation(int dictionaryID, String word);
+    public boolean hasPronuncation(long dictionaryID, String word);
 
     /**
      * Get a valid identifier for a new Corpus
@@ -85,84 +85,84 @@ public interface RiddlerRemote {
      * @param collectDate date with which this corpus should be associated (e.g. date when data collection
      * began or ended)
      */
-    public int createCorpus(int dictionaryID, MetadataWrapper metadata, Date collectDate);
+    public long createCorpus(long dictionaryID, MetadataWrapper metadata, Date collectDate);
 
     /**
      * Deeply create a new, empty Item record associated with the given Corpus.
      * Also creates corresponding Audio and Text records, each consisting of a single
      * RegionOfAudio and RegionOfText, respectively.
      * @param corpusId Corpus to which the Item should be added
-     * @return an int with valid deep linkages to its newly-created component records
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
      */
-    public int createItem(int corpusId);
+    public long createItem(long corpusId);
 
     /**
      * Deeply create a new Item record with exactly one Audio record containing one RegionOfAudio.
-     * @return an int with valid deep linkages to its newly-created component records
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
      * @param channelCount number of audio channels
      * @param data audio data itself
      */
-    public int createItemWithShortAudio(int corpusId, int samplesPerSecond, int channelCount, short[] data);
+    public long createItemWithShortAudio(long corpusId, int samplesPerSecond, int channelCount, short[] data);
 
     /**
      * Deeply create a new Item record with exactly one Audio record containing one RegionOfAudio.     
-     * @return an int with valid deep linkages to its newly-created component records
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
      * @param channelCount number of audio channels
      * @param data audio data itself
      */
-    public int createItemWithByteAudio(int corpusId, int samplesPerSecond, int channelCount, byte[] data);
+    public long createItemWithByteAudio(long corpusId, int samplesPerSecond, int channelCount, byte[] data);
 
     /**
      * Deeply create a new Item record with exactly one Audio record containing one RegionOfAudio.
-     * @return an int with valid deep linkages to its newly-created component records
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
      * @param channelCount number of audio channels
      * @param data audio data itself
      */
-    public int createItemWithIntAudio(int corpusId, int samplesPerSecond, int channelCount, int[] data);
+    public long createItemWithIntAudio(long corpusId, int samplesPerSecond, int channelCount, int[] data);
 
     /**
      * Deeply create a new Item record with exactly one Audio record containing one RegionOfAudio.
-     * @return an int with valid deep linkages to its newly-created component records
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
      * @param channelCount number of audio channels
      * @param data audio data itself
      */
-    public int createItemWithLongAudio(int corpusId, int samplesPerSecond, int channelCount, long[] data);
+    public long createItemWithLongAudio(long corpusId, int samplesPerSecond, int channelCount, long[] data);
 
     /**
      * Deeply create a new Item record with exactly one Audio record containing one RegionOfAudio.
-     * @return an int with valid deep linkages to its newly-created component records
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
      * @param channelCount number of audio channels
      * @param data audio data itself
      */
-    public int createItemWithFloatAudio(int corpusId, int samplesPerSecond, int channelCount, float[] data);
+    public long createItemWithFloatAudio(long corpusId, int samplesPerSecond, int channelCount, float[] data);
 
     /**
      * Deeply create a new Item record with exactly one Text record containing one RegionOfText.
-     * @return an int with valid deep linkages to its newly-created component records
-     * @throws RemoteException if the descriptor contains a word not found in the
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
+     * @throws RemoteException if the words list contains a word not found in the
      * provided Corpus' Dictionary
      * @param corpusId Corpus to which the Item should be added
      * @param words text with which the Item should be created; note that a List is used
      * because not all languages use spaces consistently as word delimiters
      */
-    public int createItemWithText(int corpusId, ArrayList<String> words);
+    public long createItemWithText(long corpusId, ArrayList<String> words);
 
     /**
      * Deeply create a new Item record with one Audio record and one Text record.  The Audio
      * record contains one RegionOfAudio that points to the Text record's single RegionOfText.<p/>
      * This method should be used to indicate that the Text is a transcript of the Audio.
-     * @return an int with valid deep linkages to its newly-created component records
-     * @throws RemoteException if the descriptor contains a word not found in the
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
+     * @throws RemoteException if the words list contains a word not found in the
      * provided Corpus' Dictionary
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
@@ -170,14 +170,14 @@ public interface RiddlerRemote {
      * @param data audio data itself
      * @param words transcript of the audio being sent
      */
-    public int createItemWithShortAudioAndText(int corpusId, int samplesPerSecond, int channelCount, short[] data, ArrayList<String> words);
+    public long createItemWithShortAudioAndText(long corpusId, int samplesPerSecond, int channelCount, short[] data, ArrayList<String> words);
 
     /**
      * Deeply create a new Item record with one Audio record and one Text record.  The Audio
      * record contains one RegionOfAudio that points to the Text record's single RegionOfText.<p/>
      * This method should be used to indicate that the Text is a transcript of the Audio.
-     * @return an int with valid deep linkages to its newly-created component records
-     * @throws RemoteException if the descriptor contains a word not found in the
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
+     * @throws RemoteException if the words list contains a word not found in the
      * provided Corpus' Dictionary
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
@@ -185,14 +185,14 @@ public interface RiddlerRemote {
      * @param data audio data itself
      * @param words transcript of the audio being sent
      */
-    public int createItemWithByteAudioAndText(int corpusId,  int samplesPerSecond, int channelCount, byte[] data, ArrayList<String> words);
+    public long createItemWithByteAudioAndText(long corpusId,  int samplesPerSecond, int channelCount, byte[] data, ArrayList<String> words);
 
     /**
      * Deeply create a new Item record with one Audio record and one Text record.  The Audio
      * record contains one RegionOfAudio that points to the Text record's single RegionOfText.<p/>
      * This method should be used to indicate that the Text is a transcript of the Audio.
-     * @return an int with valid deep linkages to its newly-created component records
-     * @throws RemoteException if the descriptor contains a word not found in the
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
+     * @throws RemoteException if the words list contains a word not found in the
      * provided Corpus' Dictionary
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
@@ -200,14 +200,14 @@ public interface RiddlerRemote {
      * @param data audio data itself
      * @param words transcript of the audio being sent
      */
-    public int createItemWithIntAudioAndText(int corpusId,  int samplesPerSecond, int channelCount, int[] data, ArrayList<String> words);
+    public long createItemWithIntAudioAndText(long corpusId,  int samplesPerSecond, int channelCount, int[] data, ArrayList<String> words);
 
     /**
      * Deeply create a new Item record with one Audio record and one Text record.  The Audio
      * record contains one RegionOfAudio that points to the Text record's single RegionOfText.<p/>
      * This method should be used to indicate that the Text is a transcript of the Audio.
-     * @return an int with valid deep linkages to its newly-created component records
-     * @throws RemoteException if the descriptor contains a word not found in the
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
+     * @throws RemoteException if the words list contains a word not found in the
      * provided Corpus' Dictionary
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
@@ -215,14 +215,14 @@ public interface RiddlerRemote {
      * @param data audio data itself
      * @param words transcript of the audio being sent
      */
-    public int createItemWithLongAudioAndText(int corpusId,  int samplesPerSecond, int channelCount, long[] data, ArrayList<String> words);
+    public long createItemWithLongAudioAndText(long corpusId,  int samplesPerSecond, int channelCount, long[] data, ArrayList<String> words);
 
     /**
      * Deeply create a new Item record with one Audio record and one Text record.  The Audio
      * record contains one RegionOfAudio that points to the Text record's single RegionOfText.<p/>
      * This method should be used to indicate that the Text is a transcript of the Audio.
-     * @return an int with valid deep linkages to its newly-created component records
-     * @throws RemoteException if the descriptor contains a word not found in the
+     * @return the ID of an Item with valid deep linkages to its newly-created component records
+     * @throws RemoteException if the words list contains a word not found in the
      * provided Corpus' Dictionary
      * @param corpusId Corpus to which the Item should be added
      * @param samplesPerSecond samples per second
@@ -230,7 +230,7 @@ public interface RiddlerRemote {
      * @param data audio data itself
      * @param words transcript of the audio being sent
      */
-    public int createItemWithFloatAudioAndText(int corpusId,  int samplesPerSecond, int channelCount, float[] data, ArrayList<String> words);
+    public long createItemWithFloatAudioAndText(long corpusId,  int samplesPerSecond, int channelCount, float[] data, ArrayList<String> words);
 
     /**
      * Add a RegionOfText to the given Item
@@ -238,9 +238,9 @@ public interface RiddlerRemote {
      * @param itemID Item record to which the new RegionOfText should be added
      * @param startIndex index of the word at which this text region begins
      * @param endIndex index of the word at which this text region ends
-     * @see RiddlerRemote#createItemWithText(int, java.util.ArrayList)
+     * @see RiddlerRemote#createItemWithText(long, java.util.ArrayList)
      */
-    public int createTextRegion(int itemID, int startIndex, int endIndex);
+    public long createTextRegion(long itemID, int startIndex, int endIndex);
 
     /**
      * Add a RegionOfAudio to the given Item
@@ -249,7 +249,7 @@ public interface RiddlerRemote {
      * @param beginTime time (in milliseconds) within the parent Audio record when this region begins
      * @param endTime time (in milliseconds) within the parent Audio record when this region ends
      */
-    public int createAudioRegion(int itemID, int beginTime, int endTime);
+    public long createAudioRegion(long itemID, int beginTime, int endTime);
 
     /**
      * Add a RegionOfAudio to the given Item, associating it with the given RegionOfText
@@ -260,7 +260,7 @@ public interface RiddlerRemote {
      * @param startIndex index of the word at which this text region begins
      * @param endIndex index of the word at which this text region ends
      */
-    public int createAudioRegionWithText(int itemID, int beginTime, int endTime, int startIndex, int endIndex);
+    public long createAudioRegionWithText(long itemID, int beginTime, int endTime, int startIndex, int endIndex);
 
     /**
      * Link a pre-existing audio region to a pre-existing text region
@@ -268,5 +268,5 @@ public interface RiddlerRemote {
      * @param textID ID of a Text record
      * @throws RemoteException if either the audio or text ID's are invalid
      */
-    public void associateAudioRegionWithText(int audioID, int textID);
+    public void associateAudioRegionWithText(long audioID, long textID);
 }
